pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "imio"
version = "0.1.0"
dependencies = [
    "numpy",
    "pillow",
]

[tool.setuptools.packages.find]
where = ["src"]
